Add a way to leave groups through the API
As we shift our login from a cookie-based system to a token-based one,
we need a way to leave a group with only an access token, supplementing
and ultimately replacing the cookie-authenticated method that exists at
the moment[1].

[1]: https://github.com/hypothesis/h/blob/d8928aa6/h/views/groups.py#L127-L137

The URL structure of this API is designed to fit into a future world
where groups are treated as first-class resources in the API, based on
the proposal @nickstenning wrote before he left Hypothesis[2].
Currently, the only user we support removing from the group is the
requesting user (represented with the shorthand value "me", rather than
a user ID), but we can expand this in the future to allow removal of
arbitrary users from groups.

[2]: https://notes.wtk.io/2017/07/24/api-scribblings

In line with the previous implementation, this endpoint returns an empty
`204` response, whether the requesting user was initially a member of
the group or not. At some point, we may need to revisit this decision:
for instance, if we decide to implement groups that are not visible to
anyone who knows their slug, then this could be used by a malicious user
as a way to check whether a group with a given slug exists.